Package Linux GUI as installable module
