+++ pyproject.toml
@@ -1,0 +1,41 @@
+[build-system]
+requires = ["setuptools>=65", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "vrchat-join-notification-with-pushover"
+version = "0.1.0"
+description = "Watch VRChat logs and send join notifications with optional Pushover pushes."
+readme = "README.md"
+requires-python = ">=3.8"
+license = {file = "LICENSE"}
+authors = [
+  {name = "yueplush"}
+]
+classifiers = [
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+  "License :: OSI Approved :: MIT License",
+  "Operating System :: POSIX :: Linux",
+  "Topic :: Games/Entertainment",
+  "Topic :: System :: Monitoring",
+]
+dependencies = []
+
+[project.optional-dependencies]
+tray = [
+  "pystray>=0.19",
+  "Pillow>=9",
+]
+
+[project.scripts]
+vrchat-join-notifier = "vrchat_join_notification.app:main"
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.setuptools.package-data]
+vrchat_join_notification = ["notification.ico"]
