+++ src/vrchat_join_notification/__init__.py
@@ -1,0 +1,5 @@
+"""VRChat Join Notification package."""
+
+from .app import main
+
+__all__ = ["main"]
