+++ src/vrchat_join_notification/notification.ico
